changes in product service and enable lazy loading

+++ NadinSoftProject/Services/ProductsService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Routing.Constraints;
 using NadinSoftProject.Data;
 using NadinSoftProject.Models;
 using NadinSoftProject.Models.Dto;
@@ -10,9 +11,9 @@ namespace NadinSoftProject.Services
         IEnumerable<Product> GetAllProducts();
         Product GetProductById(int id);
         Product GetProductByName(string productName);
-        void AddProduct(ProductDto productDto);
-        void DeleteProduct(Product product);
-        void UpdateProduct(ProductDto productDto);
+        void AddProduct(ProductDto productDto, int userId );
+        void DeleteProduct(Product product, int userId);
+        void UpdateProduct(ProductDto productDto, int userId);
     }
     public class ProductsService : IProductsService
     {
@@ -35,19 +36,33 @@ namespace NadinSoftProject.Services
         {
             return _db.Products.FirstOrDefault(product => product.ProductName == productName);
         }
-        public void AddProduct(ProductDto productDto)
+        public void AddProduct(ProductDto productDto, int userId)
         {
-            _db.Products.Add(_mapper.Map<Product>(productDto));
+            var user = _db.Users.FirstOrDefault(u => u.Id == userId);
+            var product = _mapper.Map<Product>(productDto);
+            product.User = user;
+            _db.Products.Add(product);
             _db.SaveChanges();
         }
-        public void DeleteProduct(Product product)
+        public void DeleteProduct(Product product, int userId)
         {
+            var user = _db.Users.FirstOrDefault(u => u.Id == userId);
+            if(user.Id != product.User.Id)
+            {
+                throw new UnauthorizedAccessException();
+            }
             _db.Products.Remove(product);
             _db.SaveChanges();
         }
-        public void UpdateProduct(ProductDto productDto)
+        public void UpdateProduct(ProductDto productDto, int userId)
         {
-            _db.Products.Update(_mapper.Map<Product>(productDto));
+            var user = _db.Users.FirstOrDefault(u => u.Id == userId);
+            var product = _mapper.Map<Product>(productDto);
+            if (user.Id != product.User.Id)
+            {
+                throw new UnauthorizedAccessException();
+            }
+            _db.Products.Update(product);
             _db.SaveChanges();
         }
     }
